+++ app/src/main/java/org/android/go/sopt/MainActivity.kt
@@ -69,6 +69,4 @@ class MainActivity : AppCompatActivity() {
 
 
 
-
-
 }
