+++ app/src/main/res/layout/header_github_repo.xml
@@ -7,6 +7,10 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="소현의 레포지터리" />
+        android:text="@string/header_pop_playlist"
+        android:layout_marginStart="10dp"
+        android:textAlignment="viewStart"
+        android:textSize="16sp"
+        android:textStyle="bold" />
 
 </LinearLayout>
