+++ app/src/main/res/values/strings.xml
@@ -8,6 +8,8 @@
     <string name="signup_pw">PW</string>
     <string name="signup_name">이름</string>
     <string name="signup_specialty">특기</string>
+    <string name="header_pop_playlist">PLAYLIST</string>
+
 
 
 </resources>
